Use async-compute
Synchronous updates at a fixed rate create a drop in framerate each
time they trigger.
Moreover, given how bevy_spatial works, there is control of when the
spatial data is updated relative to the boid updates.

So switch to async-compute, which allows the updates to run over
multiple frames. This also requires to manually control the spatial
updates through direct used of the kd-tree, instead of using
bevy_spatial